chapter10/ML-API/requirements.txt
google-cloud-storage
google-cloud-vision